Move the Push await to rule
