+++ app/src/androidTest/java/com/jraska/github/client/xpush/PushIntegrationTest.kt
@@ -1,23 +1,15 @@
 package com.jraska.github.client.xpush
 
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.IdlingRegistry
-import androidx.test.espresso.IdlingResource
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.idling.CountingIdlingResource
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.platform.app.InstrumentationRegistry
 import com.google.firebase.iid.FirebaseInstanceId
-import com.google.firebase.messaging.RemoteMessage
-import com.jraska.github.client.DecoratedServiceModelFactory
 import com.jraska.github.client.DeepLinkLaunchTest
-import com.jraska.github.client.TestUITestApp
-import com.jraska.github.client.core.android.ServiceModel
-import com.jraska.github.client.push.PushHandleModel
-import org.junit.After
 import org.junit.Assume
 import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
 
 class PushIntegrationTest {
@@ -25,25 +17,13 @@ class PushIntegrationTest {
   lateinit var pushClient: PushServerClient
   lateinit var thisDeviceToken: String
 
-  @Suppress("UNCHECKED_CAST") // We want to fail on unchecked casting
+  @get:Rule
+  val pushRule = PushAwaitRule()
+
   @Before
   fun setUp() {
     pushClient = PushServerClient.create(apiKey())
     thisDeviceToken = FirebaseInstanceId.getInstance().token!!
-
-    TestUITestApp.get().decoratedServiceFactory.decorator = object : DecoratedServiceModelFactory.Decorator {
-      override fun <T : ServiceModel> create(modelClass: Class<T>, productionFactory: ServiceModel.Factory): T {
-        return TestPushHandleModel(productionFactory.create(modelClass) as PushHandleModel) as T
-      }
-    }
-
-    IdlingRegistry.getInstance().register(PushAwaitIdlingResource.idlingResource())
-  }
-
-  @After
-  fun tearDown() {
-    IdlingRegistry.getInstance().unregister(PushAwaitIdlingResource.idlingResource())
-    TestUITestApp.get().decoratedServiceFactory.decorator = null
   }
 
   @Test
@@ -84,25 +64,7 @@ class PushIntegrationTest {
   }
 
   private fun awaitPush() {
-    PushAwaitIdlingResource.waitForPush()
+    pushRule.waitForPush()
   }
 
-  object PushAwaitIdlingResource {
-    private val countingIdlingResource = CountingIdlingResource("Push Await")
-
-    fun idlingResource(): IdlingResource = countingIdlingResource
-
-    fun waitForPush() = countingIdlingResource.increment()
-
-    fun onPush() = countingIdlingResource.decrement()
-  }
-
-  class TestPushHandleModel(
-    val productionModel: PushHandleModel,
-  ) : PushHandleModel by productionModel {
-    override fun onMessageReceived(message: RemoteMessage) {
-      productionModel.onMessageReceived(message)
-      PushAwaitIdlingResource.onPush()
-    }
-  }
 }
